Himos App and Test Update (#4156)
* Updated himos to new application format

* Added new gtests for himos

* Removed old himos tests

* Addressed excess white spaceing

* Added static declaration on app helper functions

* Removed unnecessary static declaration
