+++ isis/src/mro/apps/himos/himos.h
@@ -1,0 +1,10 @@
+#ifndef himos_h
+#define himos_h
+
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void himos(UserInterface &ui);
+}
+
+#endif
